solc abi encoding WIP
